myproject/requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
automium==0.2.6
automium-web==0.1.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
deprecation==2.1.0
Django==5.0.6
django-filter==24.2
django-keycloak==0.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
exceptiongroup==1.2.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
josepy==1.14.0
jwcrypto==1.5.6
legacy==0.1.7
MarkupSafe==2.1.5
mozilla-django-oidc==4.0.1
myapp==0.1.dev0
outcome==1.3.0.post0
packaging==24.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
PySocks==1.7.1
python-jose==3.3.0
python-keycloak==4.1.0
python-keycloak-client==0.2.3
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
urllib3==2.2.1
views==0.3
Werkzeug==3.0.3
wsproto==1.2.0
